// src/app/shared/model/VendorBankAcc.Model.ts
export class VendorBankAccObj {

    VendorBankAccId: number;
    VendorId: number;
    VendorCode: string;
    VendorEmpId: number;
    VendorEmpNo: string;
    RefBankId: number;
    BankName: string;
    BankCode: string;
    BankAccountNo: string;
    BankAccountName: string;
    IsDefault: boolean;
    RefBankid: number;
    BankBranch: string;
    Notes: string;

    constructor() {

    }
}